backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benx-quant-backend"
version = "0.1.0"
description = "FastAPI backend for Benx Quant Trading Platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "dnspython>=2.7.0",
  "motor>=3.7.0",
  "pydantic-settings>=2.7.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.28.1",
  "redis>=5.2.1",
  "python-socketio>=5.12.1",
  "email-validator>=2.2.0",
  "joblib>=1.4.2",
  "pandas>=2.2.3",
  "numpy>=2.2.3",
  "scikit-learn>=1.6.1",
  "torch>=2.6.0",
  "ta>=0.11.0",
  "orjson>=3.10.15",
  "python-multipart>=0.0.20",
  "websockets>=15.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "ai_models*", "trading_engine*"]
